+++ framework/db/mssql/QueryBuilder.php
@@ -174,35 +174,40 @@ class QueryBuilder extends \yii\db\QueryBuilder
      */
     public function alterColumn($table, $column, $type)
     {
-        $sqlAfter = [];
+        $sqlAfter = [$this->dropConstraintsForColumn($table, $column, 'D')];
 
         $columnName = $this->db->quoteColumnName($column);
         $tableName = $this->db->quoteTableName($table);
-
         $constraintBase = preg_replace('/[^a-z0-9_]/i', '', $table . '_' . $column);
 
-        $type = $this->getColumnType($type);
+        if ($type instanceof \yii\db\mssql\ColumnSchemaBuilder) {
+            $type->setAlterColumnFormat();
 
-        if (preg_match('/\s+DEFAULT\s+(["\']?\w*["\']?)/i', $type, $matches)) {
-            $type = preg_replace('/\s+DEFAULT\s+(["\']?\w*["\']?)/i', '', $type);
-            $sqlAfter[] = $this->dropConstraintsForColumn($table, $column, 'D');
-            $sqlAfter[] = $this->addDefaultValue("DF_{$constraintBase}", $table, $column, $matches[1]);
-        } else {
-            $sqlAfter[] = $this->dropConstraintsForColumn($table, $column, 'D');
+
+            $defaultValue = $type->getDefaultValue();
+            if ($defaultValue !== null) {
+                $sqlAfter[] = $this->addDefaultValue(
+                    "DF_{$constraintBase}",
+                    $table,
+                    $column,
+                    $defaultValue instanceof Expression ? $defaultValue : new Expression($defaultValue)
+                );
+            }
+
+            $checkValue = $type->getCheckValue();
+            if ($checkValue !== null) {
+                $sqlAfter[] = "ALTER TABLE {$tableName} ADD CONSTRAINT " .
+                    $this->db->quoteColumnName("CK_{$constraintBase}") .
+                    " CHECK (" . ($defaultValue instanceof Expression ?  $checkValue : new Expression($checkValue)) . ")";
+            }
+
+            if ($type->isUnique()) {
+                $sqlAfter[] = "ALTER TABLE {$tableName} ADD CONSTRAINT " . $this->db->quoteColumnName("UQ_{$constraintBase}") . " UNIQUE ({$columnName})";
+            }
         }
 
-        if (preg_match('/\s+CHECK\s+\((.+)\)/i', $type, $matches)) {
-            $type = preg_replace('/\s+CHECK\s+\((.+)\)/i', '', $type);
-            $sqlAfter[] = "ALTER TABLE {$tableName} ADD CONSTRAINT " . $this->db->quoteColumnName("CK_{$constraintBase}") . " CHECK ({$matches[1]})";
-        }
-
-        $type = preg_replace('/\s+UNIQUE/i', '', $type, -1, $count);
-        if ($count) {
-            $sqlAfter[] = "ALTER TABLE {$tableName} ADD CONSTRAINT " . $this->db->quoteColumnName("UQ_{$constraintBase}") . " UNIQUE ({$columnName})";
-        }
-
-        return 'ALTER TABLE ' . $this->db->quoteTableName($table) . ' ALTER COLUMN '
-            . $this->db->quoteColumnName($column) . ' '
+        return 'ALTER TABLE ' . $tableName . ' ALTER COLUMN '
+            . $columnName . ' '
             . $this->getColumnType($type) . "\n"
             . implode("\n", $sqlAfter);
     }
@@ -661,5 +666,4 @@ END";
         return $this->dropConstraintsForColumn($table, $column) . "\nALTER TABLE " . $this->db->quoteTableName($table)
             . " DROP COLUMN " . $this->db->quoteColumnName($column);
     }
-
 }