+++ framework/db/mssql/Schema.php
@@ -804,4 +804,12 @@ SQL;
 
         return $result;
     }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function createColumnSchemaBuilder($type, $length = null)
+    {
+        return new ColumnSchemaBuilder($type, $length, $this->db);
+    }
 }